Fixed all warnings in poplibs and enabled Werror
Summary: Fixes T10975

Test Plan: n/a

Reviewers: tomm, oleksiik, richard, Eddy

Reviewed By: Eddy

Subscribers: davidb

Maniphest Tasks: T10975

Differential Revision: https://phabricator.sourcevertex.net/D15932

+++ tools/cast_to_gfloat.hpp
@@ -65,7 +65,7 @@ RoundType convertStringToRoundType(const std::string &roundMode, Type inType,
     return RoundType::RD;
   } else if (roundMode == "SR") {
     bool isExtendedSr =
-        srBits < ((inType == FLOAT) ? manSizeFp32 : manSizeFp16);
+        srBits < unsigned((inType == FLOAT) ? manSizeFp32 : manSizeFp16);
     if (isExtendedSr) {
       return RoundType::SX;
     } else {
